Update UnitTest to print errors
The Unit Test will now print errors if encountered, instead of silently dropping them. Additionally, the unit test will NOT show errors due to not-yet implemented functions, so that students are not hit with a wall of errors at the start. This is done by ignoring errors that come from trying to call functions with arguments that haven't been implemented yet. Depends on the preexisting assumption that all function definitions exist.