Update UnitTest: Don't print err if already known
The Unit test will not print an error if the current function (unit) is known to contain an error. This avoids giving students a wall of errors for a single mistake, which is caused by the fact that the unit test will run each unit multiple times with various test cases.